Move example files for memfs to own folder
The strings for these files were taking up a lot of space, making browsing memfs.ts difficult
